Refactor for stability.
Keep track of when the update is finished, and let caller know. Implemented
with Node.js' built-in Promises. Requires Node.js 0.12 or later.

Avoid keeping internal global state, which is overwritten if module is called
with different options before the first call has finished. Now it is possible
to initiate calls for updating several records at once, collecting their
results with Promise.all().

Resolves michaelkourlas/node-cloudflare-dynamic-dns#1